Introduced a complete redesign and implementation of the Mesos CLI.
This commit introduces the basic infrastructure for the new CLI tool.
It is written in Python and is based on the plugable architecture
described in MESOS-5676.

One of the salient features of the CLI is that it can be ran both in
python directly, as well as bundled into a standalone executable that
can be deployed with no external dependencies.